Rename package and add a REPL example
- Rename package from 'algos' to 'algos_contrib' to avoid name conflict
  with the default package name in Machine Learning Toolkit (as noted in
  the official documentation).
- Add a Python REPL example for running a simple test on an example
  algorithm.
