+++ src/bin/algos_contrib/tests/test_example_algo.py
@@ -1,11 +1,13 @@
-from algos.ExampleAlgo import ExampleAlgo
+from algos_contrib.ExampleAlgo import ExampleAlgo
 from test.util import (
     assert_registered,
     assert_signatures,
 )
 
+
 def test_signatures():
     assert_signatures(ExampleAlgo)
+
 
 def test_registration():
     assert_registered(ExampleAlgo)
