+++ src/default/algos.conf
@@ -1,5 +1,5 @@
 # Here is where algorithms are registered.
 [default]
-package=algos
+package=algos_contrib
 
 [ExampleAlgo]
